have modified posts logic

+++ src/post/post.service.ts
@@ -1,8 +1,15 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreatePostDto } from './dto/create-post.dto';
 import { UpdatePostDto } from './dto/update-post.dto';
 import { PrismaService } from '../prisma/prisma.service';
 import { VotingService } from '../voting/voting.service';
+import { PageOptionsDto } from '../common/dto/page-options.dto';
+import { PageMetaDto } from '../common/dto/page-meta.dto';
+import { PageDto } from '../common/dto/page.dto';
 
 //TODO: Создай пагинацию и фильтрацию, плюс внедри счётчик голосов.
 @Injectable()
@@ -12,18 +19,34 @@ export class PostService {
     private readonly votingService: VotingService,
   ) {}
 
-  async create(createPostDto: CreatePostDto) {
+  async create(createPostDto: CreatePostDto, userId: number) {
     try {
       return await this.prisma.post.create({
-        data: createPostDto,
+        data: { ...createPostDto, userId },
       });
     } catch (error) {
       throw new Error('Error in creating post: ' + error.message);
     }
   }
 
-  async findAll() {
-    await this.prisma.post.findMany();
+  async findAll(pageOptionsDto: PageOptionsDto, categoryId?: number) {
+    const sortOrder = pageOptionsDto.order === 'ASC' ? 'asc' : 'desc';
+    const whereCondition = categoryId ? { categoryId } : {};
+    const posts = await this.prisma.post.findMany({
+      where: whereCondition,
+      skip: pageOptionsDto.skip,
+      take: pageOptionsDto.take,
+      orderBy: { createdAt: sortOrder },
+    });
+    const postsCount = await this.prisma.post.count({
+      where: whereCondition,
+    });
+    const pageMetaDto = new PageMetaDto({
+      itemCount: postsCount,
+      pageOptionsDto,
+    });
+
+    return new PageDto(posts, pageMetaDto);
   }
 
   async findOne(id: number) {
@@ -36,27 +59,30 @@ export class PostService {
     return post;
   }
 
-  async update(id: number, updatePostDto: UpdatePostDto) {
+  async update(id: number, updatePostDto: UpdatePostDto, userId: number) {
     try {
       return await this.prisma.post.update({
         where: { id },
-        data: updatePostDto,
+        data: { ...updatePostDto, userId },
       });
     } catch (error) {
       throw new NotFoundException(`Post with ID ${id} not found`);
     }
   }
 
-  async remove(id: number) {
-    try {
-      await this.prisma.post.delete({
-        where: { id },
-      });
-      return `Post with ID ${id} was deleted`;
-    } catch (error) {
-      throw new NotFoundException(
-        `Post with ID ${id} not found or could not be deleted`,
+  async remove(id: number, userId: number) {
+    const post = await this.findOne(id);
+
+    if (post.userId !== userId) {
+      throw new ForbiddenException(
+        'You do not have permission to delete this post',
       );
     }
+
+    await this.prisma.post.delete({
+      where: { id },
+    });
+
+    return `Post with ID ${id} was deleted`;
   }
 }
